built-in bundles in workerd config
PR #413 adds wrapped binding, but the intent is to enable
those only to builtin modules. However there's no easy way to
extend builtins in already built workerd.

This change adds builtins field to worker configuration with
the list of bundles to provide. These modules will be registered
(but not initialized) during compileModules.

Another usage for builtins could be private bindings/modules
with full sources not in workerd, but still available in local
development through the use of builtins.